requirements.txt
fpdf2
rich
pytest
pyfiglet